add screenshot, source code cleanup
