+++ src/packages/designer/components/Nodes/Designer/CloudNode.vue
@@ -21,333 +21,25 @@
 
 <script setup lang="ts">
 
-import { Position, Node, GraphNode, ClusterComponentTypes, useVueFlow } from "../../../index";
+import { Position, GraphNode, ClusterComponentTypes, useVueFlow } from "../../../index";
 import Handle from "../../Handle/Handle.vue";
 import type { NodeProps } from "../../../types/node";
 import { useCloudHelper } from "../../../composables";
-import { CloudType, getDefaultETCDClusterData, getDefaultLoadBalancerData, getDefaultMasterData, getDefaultRegistryData, getDefaultStorageClusterData, getDefaultStorageServerData, getDefaultWorkerData, WorkerRoles, XYPosition } from "../../../types";
-import { Cluster } from "cluster";
 
-const { onNodeDragStop, instance, store } = useVueFlow();
 const props = defineProps<NodeProps>()
-const node = store.getNode(props.id)
-const allowTypes = []
-const ids = new Map<ClusterComponentTypes, number>([[ClusterComponentTypes.LoadBalancer, 0], [ClusterComponentTypes.ETCDCluster, 0]])
-let isMasterHA = false;
+const { onNodeDragStop, instance, store } = useVueFlow();
 
-const { masters, workers, haProxy, etcdCluster, arrangeMembers } = useCloudHelper(store, props.id)
-
-watch(() => props.data.masterCount, (newVal, oldVal) => {
-    if (newVal < 1) {
-        // TODO: Message 처리
-        alert('Master 갯수 확인')
-        props.data.masterCount = oldVal;
-        return;
-    }
-
-    processNodes(ClusterComponentTypes.Master, oldVal, true)
-    processNodes(ClusterComponentTypes.Master, newVal)
-
-    nextTick(() => { arrangeMembers() })
-})
-
-watch(() => props.data.useExternalETCD, (newVal, oldVal) => {
-    if (oldVal === newVal) return;
-
-    masters.value.forEach(m => {
-        m.data.hasETCD = !newVal
-    })
-
-    processNodes(ClusterComponentTypes.ETCDCluster, 1, !newVal)
-
-    nextTick(() => { arrangeMembers() })
-})
-
-watch(() => props.data.workerCount, (newVal, oldVal) => {
-    if (oldVal === newVal) return;
-
-    processNodes(ClusterComponentTypes.Worker, oldVal, true)
-    processNodes(ClusterComponentTypes.Worker, newVal)
-
-    nextTick(() => { arrangeMembers() })
-})
-
-const getNodeData = (type: ClusterComponentTypes, label: string, workerRole?: WorkerRoles) => {
-    let data = {} as any;
-
-    switch (type) {
-        case ClusterComponentTypes.Master:
-            data = getDefaultMasterData()
-            break;
-        case ClusterComponentTypes.Worker:
-            if (workerRole) {
-                data = getDefaultWorkerData({ workerRole: workerRole })
-            } else {
-                data = getDefaultWorkerData({ workerRole: node.data.cloudType === CloudType.Openstack ? WorkerRoles.Compute : WorkerRoles.Worker })
-            }
-            break;
-        case ClusterComponentTypes.LoadBalancer:
-            data = getDefaultLoadBalancerData()
-            break;
-        case ClusterComponentTypes.Registry:
-            data = getDefaultRegistryData()
-            break;
-        case ClusterComponentTypes.StorageServer:
-            data = getDefaultStorageServerData()
-            break;
-        case ClusterComponentTypes.StorageCluster:
-            data = getDefaultStorageClusterData()
-            break;
-        case ClusterComponentTypes.ETCDCluster:
-            data = getDefaultETCDClusterData()
-            break;
-    }
-
-    data.name = label;
-    return data
-}
-
-const getNodePosition = (type: ClusterComponentTypes, seq: number) => {
-    // 실제 Cluster내의 Member Nodes 정렬은 arrangeMembers를 통해서 처리된다.
-    //const basePos = { x: 20, y: 40 }
-
-    // if (allowTypes.includes(type)) {
-    //     // Drop 생성 기본 포지션
-    //     return basePos
-    // } else {
-    //     // MasterHA 여부에 따른 LB 위치 조정
-    //     const width = 150;
-    //     const yGap = isMasterHA ? 270 : 170;
-    //     const xGap = isMasterHA ? 150 : 0
-    //     if (type === ClusterComponentTypes.Master) {
-    //         return {
-    //             x: xGap + basePos.x + (width * seq),
-    //             y: basePos.y
-    //         }
-    //     } else {
-    //         return {
-    //             x: basePos.x + (width * seq),
-    //             y: basePos.y + yGap
-    //         }
-    //     }
-    // }
-
-    return { x: 20, y: 40 }
-}
-
-const getLabel = (type: ClusterComponentTypes) => {
-    if (type === ClusterComponentTypes.Worker && node.data.cloudType === CloudType.Openstack) {
-        const name = WorkerRoles.Compute as string
-        return name.charAt(0).toUpperCase() + name.slice(1)
-    }
-    return type.charAt(0).toUpperCase() + type.slice(1)
-}
-
-const getIdLabel = (type: ClusterComponentTypes, seq: number) => {
-    const nodeName = getLabel(type)
-    let id = ""
-    let label = ""
-
-    if (ids.has(type)) {
-        const uid = ids.get(type)
-        id = `${nodeName}_${uid}`
-        ids.set(type, uid + 1)
-        label = nodeName
-    } else {
-        id = `${nodeName}_${seq}`
-        label = `${nodeName} Node #${seq}`
-    }
-
-    return { id, label }
-}
-
-const removeNodes = (type: ClusterComponentTypes) => {
-    const nodes = []
-
-    switch (type) {
-        // MasterHA인 경우 LB, Master
-        case ClusterComponentTypes.Master:
-            if (isMasterHA) {
-                nodes.push(...store.nodes.filter(item => item.parentNode === node.id && item.type === ClusterComponentTypes.LoadBalancer).map(item => ({ id: item.id, type: 'remove' })))
-            }
-            nodes.push(...store.nodes.filter(item => item.parentNode === node.id && item.type === type).map(item => ({ id: item.id, type: 'remove' })));
-            break;
-        // Worker 삭제
-        case ClusterComponentTypes.Worker:
-            nodes.push(...store.nodes.filter(item => item.parentNode === node.id && item.type === type).map(item => ({ id: item.id, type: 'remove' })));
-            break;
-        // ETCD Cluster 삭제
-        case ClusterComponentTypes.ETCDCluster:
-            nodes.push(...store.nodes.filter(item => item.parentNode === node.id && item.type === type).map(item => ({ id: item.id, type: 'remove' })))
-            break;
-
-    }
-    // 삭제할 Node에 연결된 Edges 삭제
-    const edges = []
-    const nodeIds = nodes.map(n => n.id)
-
-    edges.push(...store.edges.filter(e => nodeIds.includes(e.source) || nodeIds.includes(e.target)).map(item => ({ id: item.id, type: 'remove' })))
-    store.hooks.edgesChange.trigger(edges)
-
-    // Node 삭제
-    nextTick(() => {
-        store.hooks.nodesChange.trigger(nodes)
-    })
-}
-
-const getNode = (id: string, label: string, type: ClusterComponentTypes, position: XYPosition, workerRole?: WorkerRoles) => ({
-    id,
-    type,
-    position,
-    label,
-    parentNode: node.id,
-    resizeParent: true,
-    data: getNodeData(type, label, workerRole),
-})
-
-const addOpenstackNodes = (nodes: any[]) => {
-    // Controller
-    if (!store.nodes.some(n => n.type === ClusterComponentTypes.Worker && n.data.workerRoles === WorkerRoles.Controller)) {
-        nodes.push(getNode('controller_0', 'Controller', ClusterComponentTypes.Worker, getNodePosition(ClusterComponentTypes.Worker, 0), WorkerRoles.Controller))
-    }
-
-    // Network
-    if (!store.nodes.some(n => n.type === ClusterComponentTypes.Worker && n.data.workerRoles === WorkerRoles.Network)) {
-        nodes.push(getNode('network_0', 'Network', ClusterComponentTypes.Worker, getNodePosition(ClusterComponentTypes.Worker, 1), WorkerRoles.Network))
-    }
-}
-
-/**
- * 지정한 유형의 노드를 지정한 갯수만큼 추가 처리
- * @param type 추가할 노드 유형
- * @param nodeCount 추가할 노드 갯수
- */
-const addNodes = (type: ClusterComponentTypes, nodeCount: number) => {
-    // Node 추가
-    const newNodes = []
-
-    // MasterHA 케이스 처리
-    if (type === ClusterComponentTypes.Master) {
-        isMasterHA = nodeCount > 1
-        if (isMasterHA) {
-            newNodes.push(getNode('haproxy_0', 'HAProxy', ClusterComponentTypes.LoadBalancer, { x: 20, y: 60 }))
-        }
-    }
-
-    // Openstack Nodes 추가
-    if (node.data.cloudType === CloudType.Openstack && type === ClusterComponentTypes.Worker) addOpenstackNodes(newNodes)
-
-    // 개별 노드 추가
-    for (let i = 0; i < nodeCount; i++) {
-        const { id, label } = getIdLabel(type, i)
-        //TODO: Master/Worker, ... 구분 처리
-        //TODO: Position 배분 및 정리
-        //TODO: Cluster Sizing (Dimension) 처리
-        newNodes.push(getNode(id, label, type, getNodePosition(type, i)))
-    }
-
-    store.addNodes(newNodes);
-}
-
-const checkEdgeExist = (source: GraphNode<any>, target: GraphNode<any>): boolean => store.edges.some(e => (e.sourceNode === source && e.targetNode === target) || (e.sourceNode === target && e.targetNode === source))
-
-const connectMasterToLB = (newEdges: any[]) => {
-    masters.value.forEach(m => {
-        const id = `${haProxy.value.id}-${m.id}`
-        if (!checkEdgeExist(m, haProxy.value)) {
-            newEdges.push({ id, source: haProxy.value.id, target: m.id, sourceHandle: 'top', targetHandle: 'bottom' })
-        }
-    })
-}
-const connectWorkerToMaster = (newEdges: any[]) => {
-    workers.value.forEach(w => {
-        const id = `${masters.value[0].id}-${w.id}`
-        if (!checkEdgeExist(w, masters[0])) {
-            newEdges.push({ id, source: masters.value[0].id, target: w.id, sourceHandle: 'bottom', targetHandle: 'top' })
-        }
-    })
-}
-const connectWorkerToLB = (newEdges: any[]) => {
-    workers.value.forEach(w => {
-        const id = `${haProxy.value.id}-${w.id}`
-        if (!checkEdgeExist(w, haProxy.value)) {
-            newEdges.push({ id, source: haProxy.value.id, target: w.id, sourceHandle: 'bottom', targetHandle: 'top' })
-        }
-    })
-}
-const connectMasterToETCD = (newEdges: any[]) => {
-    masters.value.forEach(m => {
-        const id = `${etcdCluster.value.id}-${m.id}`
-        if (!checkEdgeExist(m, etcdCluster.value)) {
-            newEdges.push({ id, source: etcdCluster.value.id, target: m.id, sourceHandle: 'right', targetHandle: 'top' })
-        }
-    })
-}
-
-const addEdges = (type: ClusterComponentTypes, nodeCount: number) => {
-    const newEdges = []
-    switch (type) {
-        case ClusterComponentTypes.Master:
-            if (isMasterHA) {
-                connectMasterToLB(newEdges)
-                connectWorkerToLB(newEdges)
-            }
-            else {
-                connectWorkerToMaster(newEdges)
-            }
-            break;
-        case ClusterComponentTypes.Worker:
-            if (isMasterHA) {
-                connectWorkerToLB(newEdges)
-            } else {
-                connectWorkerToMaster(newEdges)
-            }
-            break;
-        case ClusterComponentTypes.ETCDCluster:
-            connectMasterToETCD(newEdges)
-            break;
-    }
-
-    store.addEdges(newEdges);
-}
-
-/**
- * 지정한 노드 유형에 따라 추가/삭제의 관련 노드들 처리
- * @param type 처리할 노드 유형
- * @param nodeCount 추가할 노드 갯수
- * @param isRemove 대상 유형 노드들의 삭제 여부
- */
-const processNodes = (type: ClusterComponentTypes, nodeCount: number = 1, isRemove: boolean = false) => {
-    if (isRemove) {
-        removeNodes(type)
-    } else {
-        // 노드 추가
-        addNodes(type, nodeCount)
-        // 연결 추가
-        addEdges(type, nodeCount)
-    }
-}
+const { processNodes, arrangeMembers } = useCloudHelper(store, props.id)
 
 const onDrop = (event: DragEvent) => {
+    // TODO: Drop이 발생한 경우
     event.stopPropagation();
     if (instance.value) {
-        // const type = event.dataTransfer?.getData("application/vueflow") as ClusterComponentTypes;
-        // if (checkNodeCondition(type)) {
-        //     processNodes(type)
-        // }
-
     }
 };
 
 onNodeDragStop(({ node }) => {
-    // const nodes = getNodesInside(getNodes.value, { ...props.dimensions, x: props.computedPosition.x, y: props.computedPosition.y }, transform.value);
-    // if (nodes.some((n) => n.id === node.id && n.id !== props.id)) {
-    //   node.label = `In ${props.id}`;
-    //   node.data = { group: props.id };
-    // } else if (node.data?.group === props.id) {
-    //   node.data.group = undefined;
-    //   node.label = node.id;
-    // }
+    // TODO: 내부 멤버의 Drag 처리
 });
 
 onMounted(() => {
@@ -369,8 +61,4 @@ onBeforeUnmount(() => {
 </script>
 
 <style scoped lang="scss">
-// .cloud-container {
-//     min-width: 100px;
-//     min-height: 100px;
-// }
 </style>
